pyproject.toml
[tool.poetry]
name = "docker-composer"
version = "0.8.1"
description = "Use docker-compose from within Python"
authors = ["Micha <schollm-git@gmx.com>"]
readme = "README.md"
homepage = "https://github.com/michascholl/docker-composer"
repository = "https://github.com/michascholl/docker-composer"
license = "Apache-2.0"
packages = [
    { include = "docker_composer", from = "src" },
]
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Operating System :: OS Independent",
  "Development Status :: 4 - Beta"
]


[tool.poetry.dependencies]
python = "^3.8"
attrs = "^20.3.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
black = "^20.8b1"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
addopts = "-ra --black"

[tool.black]
line-length = 88
target-version = ['py37']

[tool.isort]
sections = ['FUTURE', 'PYTEST', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
known_pytest = ['pytest']
multi_line_output = 3
line_length = 88
include_trailing_comma = 'True'


